% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get.random.score.R
\name{Get.random.score}
\alias{Get.random.score}
\title{Title}
\usage{
Get.random.score(r, genes.dist.q, c.sign, num.rounds = 1000, full.flag = F)
}
\arguments{
\item{full.flag}{}
}
\description{
Title
}
